+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/services/UserSessionService.java
@@ -6,6 +6,7 @@ import io.tech1.framework.b2b.mongodb.security.jwt.domain.events.EventSessionAdd
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.jwt.JwtRefreshToken;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.session.SessionsValidatedTuple2;
 import io.tech1.framework.domain.base.Username;
+import io.tech1.framework.domain.exceptions.cookie.CookieRefreshTokenNotFoundException;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
@@ -24,4 +25,5 @@ public interface UserSessionService {
     DbUserSession saveUserRequestMetadata(EventSessionAddUserRequestMetadata event);
     SessionsValidatedTuple2 validate(List<DbUserSession> usersSessions);
     void deleteById(String sessionId);
+    void deleteAllExceptCurrent(DbUser user, HttpServletRequest httpServletRequest) throws CookieRefreshTokenNotFoundException;
 }