+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/services/impl/UserSessionServiceImpl.java
@@ -1,5 +1,6 @@
 package io.tech1.framework.b2b.mongodb.security.jwt.services.impl;
 
+import io.tech1.framework.b2b.mongodb.security.jwt.cookies.CookieProvider;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.db.DbUser;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.db.DbUserSession;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.events.EventSessionAddUserRequestMetadata;
@@ -10,6 +11,7 @@ import io.tech1.framework.b2b.mongodb.security.jwt.repositories.UserSessionRepos
 import io.tech1.framework.b2b.mongodb.security.jwt.services.UserSessionService;
 import io.tech1.framework.b2b.mongodb.security.jwt.utilities.SecurityJwtTokenUtility;
 import io.tech1.framework.domain.base.Username;
+import io.tech1.framework.domain.exceptions.cookie.CookieRefreshTokenNotFoundException;
 import io.tech1.framework.domain.http.requests.UserRequestMetadata;
 import io.tech1.framework.domain.tuples.Tuple2;
 import io.tech1.framework.utilities.browsers.UserAgentDetailsUtility;
@@ -37,6 +39,8 @@ public class UserSessionServiceImpl implements UserSessionService {
     private final SecurityJwtPublisher securityJwtPublisher;
     // Repositories
     private final UserSessionRepository userSessionRepository;
+    // Cookie
+    private final CookieProvider cookieProvider;
     // Utilities
     private final GeoLocationFacadeUtility geoLocationFacadeUtility;
     private final SecurityJwtTokenUtility securityJwtTokenUtility;
@@ -161,5 +165,14 @@ public class UserSessionServiceImpl implements UserSessionService {
     @Override
     public void deleteById(String sessionId) {
         this.userSessionRepository.deleteById(sessionId);
+    }
+
+    @Override
+    public void deleteAllExceptCurrent(DbUser user, HttpServletRequest httpServletRequest) throws CookieRefreshTokenNotFoundException {
+        var sessions = this.userSessionRepository.findByUsername(user.getUsername());
+        var cookieRefreshToken = this.cookieProvider.readJwtRefreshToken(httpServletRequest);
+        var currentSession = this.userSessionRepository.findByRefreshToken(cookieRefreshToken.getJwtRefreshToken());
+        sessions.removeIf(session -> session.getId().equals(currentSession.getId()));
+        this.userSessionRepository.deleteAll(sessions);
     }
 }